Refactor: for upcoming func extend(downloader, transmitter and mobile app)
