+++ .gitignore
@@ -3,3 +3,5 @@
 .DS_Store
 
 app_state.json
+
+/local
