+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "Toolkit-ReClassifier"
-version = "0.1.2"
+name = "re-sourcer"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
