+++ src/classifier/mod.rs
@@ -1,0 +1,6 @@
+// 分类器模块：管理文件分类功能
+mod models;
+mod config;
+mod handlers;
+
+pub use handlers::routes;
