+++ src/downloader/mod.rs
@@ -1,0 +1,7 @@
+// 下载器模块：管理视频/文件下载功能
+mod models;
+mod handlers;
+mod ytdlp;
+
+pub use handlers::routes;
+pub use ytdlp::{get_ytdlp_path, check_ytdlp_available, get_version};
