+++ src/main.rs
@@ -1,50 +1,60 @@
 use actix_web::{middleware, web, App, HttpServer};
-use std::env;
 
-mod models;
-mod config;
-mod handlers;
+mod classifier;
+mod downloader;
+mod uploader;
 mod static_files;
 
-use handlers::*;
 use static_files::serve_static;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
-    // 获取当前目录
-    let current_dir = env::current_dir()
-        .map(|p| p.to_string_lossy().to_string())
-        .unwrap_or_else(|_| "Unknown".to_string());
-    
+    // 打印启动信息
     println!("====================================");
-    println!("  Toolkit ReClassifier");
+    println!("  🛠️  Personal Toolkit Server");
     println!("====================================");
-    println!("Current directory: {}", current_dir);
-    println!("Starting server at: http://localhost:1234");
-    println!("Opening browser automatically...");
+    println!("📍 服务地址: http://localhost:1234");
+    println!();
+    println!("📦 可用模块:");
+    println!("  - 📁 资源分类器 /classifier/");
+    println!("  - ⬇️  视频下载器 /downloader/");
+    println!("  - 📱 文件上传器 /uploader/");
+    println!();
+
+    // 检查 yt-dlp 是否可用
+    if downloader::check_ytdlp_available() {
+        match downloader::get_version() {
+            Ok(version) => println!("✅ yt-dlp 版本: {}", version),
+            Err(e) => println!("⚠️  yt-dlp 检查失败: {}", e),
+        }
+    } else {
+        println!("⚠️  yt-dlp 未找到，下载器功能将不可用");
+    }
+
     println!("====================================");
-    
-    // 在新线程中延迟打开浏览器
+    println!("🚀 正在启动服务器...");
+    println!("🌐 正在打开浏览器...");
+    println!("====================================");
+
+    // 延迟打开浏览器
     tokio::spawn(async {
         tokio::time::sleep(tokio::time::Duration::from_millis(1500)).await;
         if let Err(e) = open::that("http://localhost:1234") {
-            eprintln!("Failed to open browser: {}", e);
-            println!("Please open your browser and navigate to: http://localhost:1234");
+            eprintln!("❌ 无法自动打开浏览器: {}", e);
+            println!("请手动访问: http://localhost:1234");
         }
     });
-    
+
     HttpServer::new(|| {
         App::new()
             .wrap(middleware::Logger::default())
-            .route("/api/state", web::get().to(get_state))
-            .route("/api/folder", web::post().to(update_folder))
-            .route("/api/files", web::get().to(get_files))
-            .route("/api/move", web::post().to(move_file))
-            .route("/api/preset/save", web::post().to(save_preset))
-            .route("/api/preset/load", web::post().to(load_preset))
-            .route("/api/preset/delete", web::delete().to(delete_preset))
-            .route("/api/file/{path:.*}", web::get().to(serve_file))
-            // 使用嵌入的静态资源
+            // 分类器 API 路由
+            .service(web::scope("/api/classifier").configure(classifier::routes))
+            // 下载器 API 路由
+            .service(web::scope("/api/downloader").configure(downloader::routes))
+            // 上传器 API 路由
+            .service(web::scope("/api/uploader").configure(uploader::routes))
+            // 静态文件服务（嵌入式）
             .route("/", web::get().to(serve_static))
             .route("/{filename:.*}", web::get().to(serve_static))
     })
