+++ src/static_files.rs
@@ -8,15 +8,34 @@ pub struct Asset;
 
 // 从嵌入的资源中提供静态文件
 pub async fn serve_static(req: HttpRequest) -> Result<HttpResponse> {
-    let path = req.match_info().query("filename");
-    let path = if path.is_empty() { "index.html" } else { path };
-    
-    match Asset::get(path) {
+    let raw_path = req.match_info().query("filename");
+
+    // 构建实际路径
+    let path = if raw_path.is_empty() {
+        "index.html".to_string()
+    } else if raw_path.ends_with('/') {
+        format!("{}index.html", raw_path)
+    } else {
+        raw_path.to_string()
+    };
+
+    // 先尝试直接访问
+    let result = Asset::get(&path);
+
+    // 如果找不到，尝试添加 index.html
+    let content = if result.is_none() && !path.ends_with(".html") && !path.ends_with(".js") && !path.ends_with(".css") {
+        let index_path = format!("{}/index.html", path);
+        Asset::get(&index_path)
+    } else {
+        result
+    };
+
+    match content {
         Some(content) => {
-            let mime_type = mime_guess::from_path(path)
+            let mime_type = mime_guess::from_path(&path)
                 .first_or_octet_stream()
                 .to_string();
-            
+
             Ok(HttpResponse::Ok()
                 .content_type(mime_type)
                 .body(content.data.into_owned()))
