+++ src/uploader/mod.rs
@@ -1,0 +1,5 @@
+// 上传器模块：接收来自手机端的文件上传
+mod models;
+mod handlers;
+
+pub use handlers::routes;
